23 nisan

+++ Entities/Concrete/OrderDetail.cs
@@ -9,6 +9,6 @@ namespace Entities.Concrete
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public int Count { get; set; }
-        public decimal SalePrice { get; set; }
+        public int SalePrice { get; set; }
     }
 }
